server handlers.js manages all data from routs
